player: fix MeloTags sharing with playlist
In order to fix all issues of MeloTags instance sharing with
MeloPlaylist, a copy of MeloTags is done at play in MeloPlayer and
original is passed to MeloPlaylist.
The override cover URL option is also enabled in MeloPlaylistSimple
to prevent using cover URL from some MeloBrowser in which a donwload
of thumbnail / cover art is performed.