player: check grabber binary at init
In some cases, first download or update of grabber binary can be
interrupted and binary can be corrupted or null. In order to manage
these kind of cases, we check size of binary and its code return. If
something seems to be wrong, a download of a fresh binary image is
performed.