debian: fix python dependency for cross-build
When Debian package is cross built, the python3-dev of target
architecture is conflicting with build machine. Then, the package is
replaced with libpython3-dev to fix conflicts.
However, the python3.pc pkg-config file is not exported until bullseye
version (it was exported by python3-dev). To fix that issue, the
configure rules is overridden to create the symbolic link if
necessary.